Extra/PY/imgpdf/requirements.txt
Flask
fpdf
Pillow
